+++ src/loki/polygon_search.cc
@@ -33,26 +33,12 @@ static const auto Haversine = [] {
   return bg::strategy::distance::haversine<float>(vm::kRadEarthMeters);
 };
 
-void correct_ring(ring_bg_t& ring) {
-  // close open rings
-  bool is_open =
-      (ring.begin()->lat() != ring.rbegin()->lat() || ring.begin()->lng() != ring.rbegin()->lng());
-  if (!ring.empty() && is_open) {
-    ring.push_back(ring[0]);
-  }
-
-  // reverse ring if counter-clockwise
-  if (vm::polygon_area(ring) > 0) {
-    std::reverse(ring.begin(), ring.end());
-  }
-}
-
 ring_bg_t PBFToRing(const valhalla::Ring& ring_pbf) {
   ring_bg_t new_ring;
   for (const auto& coord : ring_pbf.coords()) {
     new_ring.push_back({coord.lng(), coord.lat()});
   }
-  correct_ring(new_ring);
+  bg::correct(new_ring);
   return new_ring;
 }